move functions to 3d modeler
