+++ pyaedt/modeler/modeler3d.py
@@ -7,6 +7,7 @@ import os.path
 import warnings
 
 # from pyaedt.generic.general_methods import property
+from pyaedt.generic.general_methods import GrpcApiError
 from pyaedt.generic.general_methods import generate_unique_name
 from pyaedt.generic.general_methods import pyaedt_function_handler
 from pyaedt.modeler.cad.Modeler import GeometryModeler
@@ -1363,3 +1364,124 @@ class Modeler3D(GeometryModeler, Primitives3D, object):
             return segment_objects, mesh_objects
         else:
             return segment_objects
+
+    @pyaedt_function_handler
+    def change_region_padding(self, padding_data, padding_type, direction=None, region_name="Region"):
+        """
+        Change region padding settings.
+
+        Parameters
+        ----------
+        padding_data : str or list of str
+            Padding value (with unit if necessary). A list of padding values must have corresponding
+            elements in ``padding_type`` and ``direction`` arguments.
+        padding_type : str or list of str
+            Padding type. Available options are ``"Percentage Offset"``, ``"Transverse Percentage Offset"``,
+            ``"Absolute Offset"``, ``"Absolute Position"``.
+        direction : str or list of str, optional
+            Direction to which apply the padding settings. A direction can be ``"+X"``, ``"-X"``,
+            `"+Y"``, ``"-Y"``, ``"+Z"`` or ``"-Z"``. Default is ``None``, in which case all the
+            directions are used (in the order written in the previous sentence).
+        region_name : str optional
+            Region name. Default is ``Region``.
+
+        Returns
+        -------
+        bool
+            ``True`` if successful, else ``None``.
+
+        Examples
+        ----------
+        >>> import pyaedt
+        >>> app = pyaedt.Icepak()
+        >>> app.change_region_padding("10mm", padding_type="Absolute Offset", direction="-X")
+        """
+        available_directions = ["+X", "-X", "+Y", "-Y", "+Z", "-Z"]
+        available_paddings = [
+            "Percentage Offset",
+            "Transverse Percentage Offset",
+            "Absolute Offset",
+            "Absolute Position",
+        ]
+        if not isinstance(padding_data, list):
+            padding_data = [padding_data]
+        if not isinstance(padding_type, list):
+            padding_type = [padding_type]
+        if direction is None:
+            direction = available_directions
+        else:
+            if not isinstance(direction, list):
+                direction = [direction]
+            if not all(dire in available_directions for dire in direction):
+                raise Exception("Check ``axes`` input.")
+        if not all(pad in available_paddings for pad in padding_type):
+            raise Exception("Check ``padding_type`` input.")
+
+        modify_props = []
+        for i in range(len(padding_data)):
+            modify_props.append(["NAME:" + direction[i] + " Padding Type", "Value:=", padding_type[i]])
+            modify_props.append(["NAME:" + direction[i] + " Padding Data", "Value:=", padding_data[i]])
+
+        try:
+            create_region_name = self.modeler.oeditor.GetChildObject(region_name).GetChildNames()[0]
+            self.modeler.oeditor.ChangeProperty(
+                list(
+                    [
+                        "NAME:AllTabs",
+                        list(
+                            [
+                                "NAME:Geometry3DCmdTab",
+                                list(["NAME:PropServers", region_name + ":" + create_region_name]),
+                                list(["NAME:ChangedProps"] + modify_props),
+                            ]
+                        ),
+                    ]
+                )
+            )
+            return True
+        except (GrpcApiError, SystemExit):
+            return False
+
+    @pyaedt_function_handler
+    def change_region_coordinate_system(self, region_cs="Global", region_name="Region"):
+        """
+        Change region coordinate system.
+
+        Parameters
+        ----------
+        region_cs : str, optional
+            Region coordinate system. Default is ``Global``.
+        region_name : str optional
+            Region name. Default is ``Region``.
+
+        Returns
+        -------
+        bool
+            ``True`` if successful, else ``None``.
+
+        Examples
+        ----------
+        >>> import pyaedt
+        >>> app = pyaedt.Icepak()
+        >>> app.modeler.create_coordinate_system(origin=[1, 1, 1], name="NewCS")
+        >>> app.change_region_coordinate_system(region_cs="NewCS")
+        """
+        try:
+            create_region_name = self.modeler.oeditor.GetChildObject(region_name).GetChildNames()[0]
+            self.modeler.oeditor.ChangeProperty(
+                list(
+                    [
+                        "NAME:AllTabs",
+                        list(
+                            [
+                                "NAME:Geometry3DCmdTab",
+                                list(["NAME:PropServers", region_name + ":" + create_region_name]),
+                                list(["NAME:ChangedProps", list(["NAME:Coordinate System", "Value:=", region_cs])]),
+                            ]
+                        ),
+                    ]
+                )
+            )
+            return True
+        except (GrpcApiError, SystemExit):
+            return False
